pyproject.toml
[project]
name = "fastapi-server"
version = "0.1.0"
description = ""
authors = [
    {name = "ChoYoungwon",email = "lka1116@naver.com"}
]
readme = "README.md"
requires-python = "==3.12.*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "==3.12.*"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
py-ulid = "^1.0.3"
sqlalchemy = "^2.0.37"
mysqlclient = "^2.2.7"
alembic = "^1.14.1"
dependency-injector = "^4.45.0"
bcrypt = "3.2.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.10.5"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.20"
python-dotenv = "^0.21.0"
pydantic-settings = "^2.7.1"
celery = "^5.4.0"
requests = "^2.32.3"
pymysql = "^1.1.1"
pytz = "^2025.1"
redis = "^5.2.1"
aiomysql = "^0.2.0"
